modify mapping

+++ src/main/java/com/qeeka/repository/BaseSearchRepository.java
@@ -76,7 +76,7 @@ public abstract class BaseSearchRepository<T> {
         //parse query group to simple query domain
         QueryModel query = queryParser.parse(queryRequest.getQueryGroup());
 
-        StringBuilder hql = new StringBuilder("FROM ").append(entityName);
+        StringBuilder hql = new StringBuilder("SELECT E FROM ").append(entityName).append(" AS E ");
         if (StringUtils.hasText(query.getStatement())) {
             hql.append(" WHERE ").append(query.getStatement());
         }
@@ -195,7 +195,12 @@ public abstract class BaseSearchRepository<T> {
     public T get(Object id) {
         StopWatch watch = new StopWatch();
         try {
-            return (T) entityManager.find(entityClass, id);
+            Object entity = entityManager.find(entityClass, id);
+            if (entity != null) {
+                return (T) entity;
+            } else {
+                return null;
+            }
         } finally {
             logger.debug("get, entityClass={}, id={}, elapsedTime={}", entityClass.getName(), id, watch.elapsedTime());
         }
@@ -263,10 +268,10 @@ public abstract class BaseSearchRepository<T> {
      * @param params
      * @return
      */
-    public int update(String queryString, Map<String, Object> params) {
+    public int update(CharSequence queryString, Map<String, Object> params) {
         StopWatch watch = new StopWatch();
         try {
-            Query query = entityManager.createQuery(queryString);
+            Query query = entityManager.createQuery(queryString.toString());
             if (params != null)
                 for (Map.Entry<String, Object> entry : params.entrySet()) {
                     query.setParameter(entry.getKey(), entry.getValue());
@@ -298,7 +303,9 @@ public abstract class BaseSearchRepository<T> {
         StopWatch watch = new StopWatch();
         try {
             T entity = get(id);
+            if (entity != null) {
             entityManager.remove(entity);
+            }
         } finally {
             logger.debug("delete, entityClass={},id={}, elapsedTime={}", entityClass.getName(), id, watch.elapsedTime());
         }
@@ -353,7 +360,7 @@ public abstract class BaseSearchRepository<T> {
      * @param queryString
      * @return
      */
-    public List<T> find(String queryString) {
+    public List<T> find(CharSequence queryString) {
         return find(queryString, null);
     }
 
@@ -364,10 +371,10 @@ public abstract class BaseSearchRepository<T> {
      * @param params
      * @return
      */
-    public List<T> find(String queryString, Map<String, Object> params) {
+    public List<T> find(CharSequence queryString, Map<String, Object> params) {
         StopWatch watch = new StopWatch();
         try {
-            Query query = entityManager.createQuery(queryString);
+            Query query = entityManager.createQuery(queryString.toString());
             if (params != null)
                 for (Map.Entry<String, Object> entry : params.entrySet()) {
                     query.setParameter(entry.getKey(), entry.getValue());
@@ -386,7 +393,7 @@ public abstract class BaseSearchRepository<T> {
      * @param fetchSize
      * @return
      */
-    public List<T> find(String queryString, int offset, int fetchSize) {
+    public List<T> find(CharSequence queryString, int offset, int fetchSize) {
         return find(queryString, null, offset, fetchSize);
     }
 
@@ -399,10 +406,10 @@ public abstract class BaseSearchRepository<T> {
      * @param fetchSize
      * @return
      */
-    public List<T> find(String queryString, Map<String, Object> params, int offset, int fetchSize) {
+    public List<T> find(CharSequence queryString, Map<String, Object> params, int offset, int fetchSize) {
         StopWatch watch = new StopWatch();
         try {
-            Query query = entityManager.createQuery(queryString);
+            Query query = entityManager.createQuery(queryString.toString());
             if (params != null) {
                 for (Map.Entry<String, Object> entry : params.entrySet()) {
                     query.setParameter(entry.getKey(), entry.getValue());
@@ -458,7 +465,7 @@ public abstract class BaseSearchRepository<T> {
      * @param queryString
      * @return
      */
-    public <X> X findUniqueResult(String queryString) {
+    public <X> X findUniqueResult(CharSequence queryString) {
         return findUniqueResult(queryString, null);
     }
 
@@ -469,10 +476,10 @@ public abstract class BaseSearchRepository<T> {
      * @param params
      * @return
      */
-    public <X> X findUniqueResult(String queryString, Map<String, Object> params) {
+    public <X> X findUniqueResult(CharSequence queryString, Map<String, Object> params) {
         StopWatch watch = new StopWatch();
         try {
-            Query query = entityManager.createQuery(queryString);
+            Query query = entityManager.createQuery(queryString.toString());
             if (params != null) {
                 for (Map.Entry<String, Object> entry : params.entrySet()) {
                     query.setParameter(entry.getKey(), entry.getValue());
@@ -493,6 +500,105 @@ public abstract class BaseSearchRepository<T> {
         return results.get(0);
     }
 
+
+    //-------------------------------------------------------
+    //                      SQL Query
+    //-------------------------------------------------------
+
+    /**
+     * find list by native query
+     *
+     * @param sql
+     * @return
+     */
+    public List<T> findByNativeQuery(CharSequence sql) {
+        return findByNativeQuery(sql);
+    }
+
+    /**
+     * find list by native query & params
+     *
+     * @param sql
+     * @param params
+     * @return
+     */
+    public List<T> findByNativeQuery(CharSequence sql, Map<String, Object> params) {
+        return findByNativeQuery(sql, params);
+    }
+
+    /**
+     * find list by native query & sql & offset & size
+     *
+     * @param sql
+     * @param offset
+     * @param size
+     * @return
+     */
+    public List<T> findByNativeQuery(CharSequence sql, int offset, int size) {
+        return findByNativeQuery(sql, offset, size);
+    }
+
+    /**
+     * find list by native query & sql & params & offset & size
+     *
+     * @param sql
+     * @param params
+     * @param offset
+     * @param size
+     * @return
+     */
+    public List<T> findByNativeQuery(CharSequence sql, Map<String, Object> params, Integer offset, Integer size) {
+        Query namedQuery = entityManager.createNativeQuery(sql.toString(), entityClass);
+        if (params != null) {
+            for (Map.Entry<String, Object> entry : params.entrySet()) {
+                namedQuery.setParameter(entry.getKey(), entry.getValue());
+            }
+        }
+        if (offset != null) {
+            namedQuery.setFirstResult(offset);
+        }
+        if (size != null) {
+            namedQuery.setMaxResults(size);
+        }
+        return namedQuery.getResultList();
+    }
+
+    /**
+     * find unique by native query & sql
+     *
+     * @param sql
+     * @param <X>
+     * @return
+     */
+    public <X> X findUniqueNativeQuery(CharSequence sql) {
+        return findUniqueNativeQuery(sql, null);
+    }
+
+    /**
+     * find unique by native query & sql & params
+     *
+     * @param sql
+     * @param params
+     * @param <X>
+     * @return
+     */
+    public <X> X findUniqueNativeQuery(CharSequence sql, Map<String, Object> params) {
+        StopWatch watch = new StopWatch();
+        try {
+            Query query = entityManager.createNativeQuery(sql.toString());
+            if (params != null) {
+                for (Map.Entry<String, Object> entry : params.entrySet()) {
+                    query.setParameter(entry.getKey(), entry.getValue());
+                }
+            }
+            List<X> results = query.getResultList();
+            return getUniqueResult(results);
+        } finally {
+            logger.debug("findUniqueNativeResult, query={}, params={}, elapsedTime={}", sql, params, watch.elapsedTime());
+        }
+    }
+
+
     private final class StopWatch {
         private long start;
 
@@ -511,3 +617,4 @@ public abstract class BaseSearchRepository<T> {
     }
 
 }
+
